Only deploy data to workers if it does not exist (#91)
* Only deploy data to workers if it does not exist

* Add switches for data overwrite and cert regeneration to deploy script

+++ scripts/deploy-data.sh
@@ -6,12 +6,31 @@ cd ..
 USERNAME=$(id -u -n)
 WORKERS=$(docker-machine ls | grep -c "worker")
 
+while [[ $# -gt 0 ]]
+do
+  arg="$1"
+
+  case $arg in
+    -o)
+    OVERWRITE="1"
+    ;;
+  esac
+  shift # past arg
+done
+
 echo "========================================================================"
 echo "  Preparing workers with data"
 echo "========================================================================"
 
 for i in $(seq 1 $WORKERS); do
   MACHINE=$USERNAME-docker-worker$i
-  echo "-> deploying data to $MACHINE"
-  docker-machine scp -r ./data $MACHINE:/home/docker/
+
+  # Only deploy data if it does not already exists on host or if override switch was specified
+  docker-machine ssh $MACHINE "ls /home/docker/data" &> /dev/null
+  if [[ $? -eq 1 ||  $OVERWRITE -eq "1" ]]; then
+    echo "-> deploying data to $MACHINE"
+    docker-machine scp -r ./data $MACHINE:/home/docker/
+  else
+    echo "Not deploying data to $MACHINE, since it already exists"
+  fi
 done
